handle gRPC err
