+++ backend/external/internal/command2Internal.go
@@ -4,6 +4,7 @@ import (
 	"context"
 	"fmt"
 	"google.golang.org/grpc"
+	"google.golang.org/grpc/status"
 	pb "ueckoken/plarail2021-soft-external/spec"
 )
 
@@ -18,11 +19,13 @@ type Command2Internal struct {
 	env     *Env
 }
 
+// NewCommand2Internal is Constructor of CommandInternal.
+// CommandInternal Struct has a method to talk to Internal server with gRPC.
 func NewCommand2Internal(state StationState, e *Env) *Command2Internal {
 	return &Command2Internal{station: &state, env: e}
 }
 
-// sendRaw is make a connection to internal server and talk with internal server.
+// sendRaw is making a connection to internal server and talk with internal server.
 // This method will return gRPC response and gRPC error val.
 // If you want join gRPC response Status Code and gRPC error msg, please use Command2Internal.trapResponseGrpcErr method.
 func (c2i *Command2Internal) sendRaw() (*pb.ResponseSync, error) {
@@ -56,22 +59,23 @@ func (c2i *Command2Internal) convert2pb() *pb.RequestSync {
 }
 
 func trapResponseGrpcErr(rs *pb.ResponseSync, grpcErr error) error {
-	if rs == nil { // gRPC error occur
+	sta, ok := status.FromError(grpcErr)
+	if (sta != nil && ok) || rs == nil { // gRPC error occur
 		return fmt.Errorf("gRPC Err: %w", grpcErr)
-	} else { // check Response Status
-		switch rs.Response.String() {
-		case UNKNOWN:
-			return fmt.Errorf("gRPC Err: %w; gRPC Response status is %s", grpcErr, UNKNOWN)
-		case SUCCESS:
-			if grpcErr != nil {
-				return fmt.Errorf("gRPC Err: %w; gRPC Response status is %s", grpcErr, SUCCESS)
-			} else {
-				return nil
-			}
-		case FAILED:
-			return fmt.Errorf("gRPC Err: %w; gRPC Response status is %s", grpcErr, FAILED)
-		default:
-			return fmt.Errorf("gRPC Err: %w; Unknown error is occured", grpcErr)
+	}
+	// check Response Status
+	switch rs.Response.String() {
+	case UNKNOWN:
+		return fmt.Errorf("gRPC Err: %w; gRPC Response status is %s", grpcErr, UNKNOWN)
+	case SUCCESS:
+		if grpcErr != nil {
+			return fmt.Errorf("gRPC Err: %w; gRPC Response status is %s", grpcErr, SUCCESS)
+		} else {
+			return nil
 		}
+	case FAILED:
+		return fmt.Errorf("gRPC Err: %w; gRPC Response status is %s", grpcErr, FAILED)
+	default:
+		return fmt.Errorf("gRPC Err: %w; Unknown error is occured", grpcErr)
 	}
 }
